+++ packages/side-nav/src/SideNavGroup/SideNavGroup.spec.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { render, cleanup, screen, fireEvent } from '@testing-library/react';
 import Icon from '@leafygreen-ui/icon';
-import { SideNavGroup, SideNavItem } from '.';
+import { SideNavGroup } from './';
 import { SideNavGroupProps } from './SideNavGroup';
+import { SideNavItem } from '../SideNavItem';
 
 describe('packages/side-nav', () => {
   const testIds = {
