+++ packages/side-nav/src/SideNavGroup/index.ts
@@ -1,0 +1,1 @@
+export { default as SideNavGroup } from './SideNavGroup';
